Split apart command_subprocess.cc
Bug: b/533572315
