+++ base/cvd/cuttlefish/process/BUILD.bazel
@@ -7,19 +7,32 @@ package(
 cf_build_test(name = "cf_build_test")
 
 cf_cc_library(
-    name = "command_subprocess",
-    srcs = ["command_subprocess.cc"],
-    hdrs = ["command_subprocess.h"],
+    name = "command",
+    srcs = ["command.cc"],
+    hdrs = ["command.h"],
     depend_on_what_you_use_enabled = False,
     deps = [
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/process:subprocess",
+        "//cuttlefish/process:subprocess_options",
         "//cuttlefish/result",
         "//libbase",
         "@abseil-cpp//absl/log",
         "@abseil-cpp//absl/log:check",
         "@abseil-cpp//absl/strings",
+    ],
+)
+
+cf_cc_library(
+    name = "command_subprocess",
+    hdrs = ["command_subprocess.h"],
+    deps = [
+        "//cuttlefish/process:command",
+        "//cuttlefish/process:execute",
+        "//cuttlefish/process:subprocess",
+        "//cuttlefish/process:subprocess_options",
     ],
 )
 
@@ -30,6 +43,20 @@ cf_cc_library(
     deps = [
         "@abseil-cpp//absl/log",
         "@abseil-cpp//absl/strings",
+    ],
+)
+
+cf_cc_library(
+    name = "execute",
+    srcs = ["execute.cc"],
+    hdrs = ["execute.h"],
+    deps = [
+        "//cuttlefish/common/libs/fs",
+        "//cuttlefish/process:command",
+        "//cuttlefish/process:subprocess",
+        "//cuttlefish/process:subprocess_options",
+        "//cuttlefish/result:expect",
+        "//cuttlefish/result:result_type",
     ],
 )
 
@@ -45,4 +72,21 @@ cf_cc_library(
         "//libbase",
         "@abseil-cpp//absl/log",
     ],
+)
+
+cf_cc_library(
+    name = "subprocess",
+    srcs = ["subprocess.cc"],
+    hdrs = ["subprocess.h"],
+    deps = [
+        "//cuttlefish/result:expect",
+        "//cuttlefish/result:result_type",
+        "@abseil-cpp//absl/log",
+    ],
+)
+
+cf_cc_library(
+    name = "subprocess_options",
+    srcs = ["subprocess_options.cc"],
+    hdrs = ["subprocess_options.h"],
 )
